refactor datafusion cli to be more modular
